[s390x] cmd/asm/internal/s390x: use MULLW for small multiplication
MULLW can be used for 32-bit multiplication when both operands are
signed. It can also be used for any multiplications (signed or
unsigned) that are less than 32-bits wide.

MULLW compiles down to a MSGFR instruction. The MSGFR instruction
sign extends one operand to 64-bits which is why it can't be used
for unsigned 32-bit multiplication.